optimize: optimize tool code for extension
