+++ tool/cmd/kitex/args/args.go
@@ -142,9 +142,12 @@ func (a *Arguments) buildFlags(version string) *flag.FlagSet {
 		"no_processor",
 	)
 
+	a.CmdExample = cmdExample
+
 	for _, e := range a.extends {
 		e.Apply(f)
 	}
+
 	f.Usage = func() {
 		fmt.Fprintf(os.Stderr, `Version %s
 Usage: %s [flags] IDL
@@ -152,9 +155,10 @@ Usage: %s [flags] IDL
 Examples:
 %s
 Flags:
-`, a.Version, os.Args[0], cmdExample)
+`, a.Version, os.Args[0], a.CmdExample)
 		f.PrintDefaults()
 	}
+
 	return f
 }
