+++ tool/cmd/kitex/main.go
@@ -17,11 +17,11 @@ package main
 import (
 	"bytes"
 	"flag"
-	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
+
+	"github.com/cloudwego/kitex/tool/cmd/kitex/utils"
 
 	"github.com/cloudwego/kitex/tool/cmd/kitex/sdk"
 
@@ -102,8 +102,6 @@ func main() {
 	}
 
 	if args.IDLType == "thrift" && args.Rapid {
-		// 「Thrift」场景，代替进程调用，如果用到了其他插件，这些插件还是走进程调用（因为先不在 Kitex 里固定写死插件版本）
-		// 相比进程唤起，少了 thriftgo bin 检查，因为不需要了
 		err = sdk.InvokeThriftgoBySDK(curpath, cmd)
 	} else {
 		err = kargs.ValidateCMD(cmd.Path, args.IDLType)
@@ -125,47 +123,5 @@ func main() {
 		os.Exit(1)
 	}
 NormalExit:
-	if args.IDLType == "thrift" {
-		cmd := "go mod edit -replace github.com/apache/thrift=github.com/apache/thrift@v0.13.0"
-		argv := strings.Split(cmd, " ")
-		err := exec.Command(argv[0], argv[1:]...).Run()
-
-		res := "Done"
-		if err != nil {
-			res = err.Error()
-		}
-		log.Warn("Adding apache/thrift@v0.13.0 to go.mod for generated code ..........", res)
-	}
-
-	// remove kitex.yaml generated from v0.4.4 which is renamed as kitex_info.yaml
-	if args.ServiceName != "" {
-		DeleteKitexYaml()
-	}
-
-	// If hessian option is java_extension, replace *java.Object to java.Object
-	if thriftgo.EnableJavaExtension(args.Config) {
-		if err = thriftgo.Hessian2PatchByReplace(args.Config, ""); err != nil {
-			log.Warn("replace java object fail, you can fix it then regenerate", err)
-		}
-	}
-}
-
-func DeleteKitexYaml() {
-	// try to read kitex.yaml
-	data, err := ioutil.ReadFile("kitex.yaml")
-	if err != nil {
-		if !os.IsNotExist(err) {
-			log.Warn("kitex.yaml, which is used to record tool info, is deprecated, it's renamed as kitex_info.yaml, you can delete it or ignore it.")
-		}
-		return
-	}
-	// if kitex.yaml exists, check content and delete it.
-	if strings.HasPrefix(string(data), "kitexinfo:") {
-		err = os.Remove("kitex.yaml")
-		if err != nil {
-			log.Warn("kitex.yaml, which is used to record tool info, is deprecated, it's renamed as kitex_info.yaml, you can delete it or ignore it.")
-		} else {
-			log.Warn("kitex.yaml, which is used to record tool info, is deprecated, it's renamed as kitex_info.yaml, so it's automatically deleted now.")
-		}
-	}
+	utils.OnKitexToolNormalExit(args)
 }
