+++ tool/internal_pkg/generator/generator.go
@@ -100,6 +100,7 @@ type Generator interface {
 
 // Config .
 type Config struct {
+	CmdExample            string
 	Verbose               bool
 	GenerateMain          bool // whether stuff in the main package should be generated
 	GenerateInvoker       bool // generate main.go with invoker when main package generate
